// web/src/components/list-filter-bar/interface.ts
/*
 *  Copyright 2026 The InfiniFlow Authors. All Rights Reserved.
 *
 *  Licensed under the Apache License, Version 2.0 (the "License");
 *  you may not use this file except in compliance with the License.
 *  You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 *  Unless required by applicable law or agreed to in writing, software
 *  distributed under the License is distributed on an "AS IS" BASIS,
 *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *  See the License for the specific language governing permissions and
 *  limitations under the License.
 */

export type FilterType = {
  id: string;
  field?: string;
  label: string | JSX.Element;
  list?: FilterType[];
  value?: string | string[];
  count?: number;
  canSearch?: boolean;
};
export type FilterCollection = {
  field: string;
  label: string;
  list: FilterType[];
  canSearch?: boolean;
};
export type FilterValue = Record<
  string,
  Array<string> | Record<string, Array<string>>
>;
export type FilterChange = (value: FilterValue) => void;